backend/requirements.txt
fastapi
uvicorn
transformers
torch
python-multipart
python-dotenv